refactor(core): move workspace modules into workspace/ directory
Replace #[path] attributes with standard directory-based module
resolution, consistent with the existing scripting/ pattern.

workspace.rs + workspace_*.rs → workspace/mod.rs + workspace/*.rs
