+++ krillnotes-core/src/core/workspace/attachments.rs
