+++ krillnotes-core/src/core/workspace/hooks.rs
