+++ krillnotes-core/src/core/workspace/scripts.rs
